Cache result of Candidate.setFiles()
Because we now call `Candidate.setFiles()` two times (both from
`Candidate.park()` and from `Candidate.handlePassedCandidate()`),
and because that method is of considerable complexity,
let’s cache the result in an instance variable, just like with
the existing instance variables `Candidate._fileName`,
`Candidate._alternative`, etc.

(Of course we could define this instance variable as well as
the complete `setFiles()` method in the subclass `FPCandidate`,
because right now it is not even possible to define a set
delist nomination.  But I add both the method and the instance
variable to the base class `Candidate` for two reasons:

1. Right now all instance variables are defined in the base class,
almost all methods reside there, including set-related stuff;
it’s just the way the classes are organized.

2. Actually set delist nominations would make sense -- of course
they would be rather rare, but sometimes people really want
to delist a whole list of files, e.g. all FPs by a photographer,
etc.  Therefore some day (I hope) we will define a simple and
clean way to create a set delist nomination and add the necessary
code to handle such a nomination to this bot program.)