Use colours in output more consistently
We use coloured output to emphasize new tasks (blue) as well
as errors (red).  This is useful because it structures
the output and helps to spot errors more easily, but it is
only really helpful if the colours are used sparingly and
consistently, else it rather confuses the user who supervises
the bot.  There are three points which can be improved:

1. Some hints saying that a specific step was skipped because
the image, template etc. is already on the page, etc., are
emphasized in red, like important errors; but there is no need
to take action, so IMHO the emphasis is rather confusing here.
Some other hints of the same kind are not emphasized by colour.

2. Some hints or warnings use yellow, but that colour is not
used consistently with a clear meaning.  Sometimes it is used
for warnings, sometimes even for errors (which normally use red).
In `main()` it is used once to emphasize that a certain task
is not implemented, but another very similar hint in the same
function does not use any colour.

3. In `notifyNominator()` and `notifyUploader()`, we have some
warnings that the user talk page cannot be modified because
it is undefined, locked, etc.  Most of these warnings say
‘but ignoring’ because ‘it’s just the user notification’
or similar.  So these are clearly just warnings, not errors.
Nevertheless some of these messages are still printed in red,
like important errors which require attention and action.

Let’s make the coloured output even more helpful by settling
these issues as follows:

1. Let’s use the colour red for all errors which require attention
and that the user takes action, and *only* for such errors.

2. Let’s use the colour yellow for mere warnings which probably
do not require any action.  This encompasses the warnings
in `notifyNominator()` and `notifyUploader()` which state
that a user talk page could not be modified because it was locked
etc., but that this issue is unimportant and therefore ignored.
It also encompasses the hints in `main()` that a certain task
is disabled or just not implemented.

3. Let’s use the colour blue only for the ‘subheadings’
which state that we begin with a new task.

4. Hints that a step was skipped because it has already been
done before should always use the default colour (= no colour)
because they don’t require any special attention or action.

Added two comments explaining why non-existent user talk pages
are unusual and therefore yield a warning (use yellow text),
unlike e.g. non-existent monthly overview pages which are just
created from scratch if they do not exist yet.