Use separate functions to print errors, warnings
After making the use of colours in the CLI output more consistent,
it’s obvious that the numerous explicit references to colour names
are a bit cumbersome and prone to typos etc.  Now we have a clear
semantic concept for the use of colours in output (errors: red,
warnings: yellow, subheadings for tasks: blue), so let’s also make
these semantics explicit by referring directly to errors and warnings
instead of hiding the semantics behind the somewhat fragile use
of explicit colour names.

We could do this by using an enumerated value as parameter
to the `out()` value, something like `mode=OutputMode.WARNING`,
`mode=OutputMode.ERROR` would work well.  But it has the little
disadvantage that it makes the lines of code even longer.

Therefore I propose to use just different functions to print errors
and warnings.  This way the code is very clear and concise:
you see at first glance at the code which kind of output is used,
the semantics are completely obvious and still the code gets
even shorter and less prone to typos and misunderstandings.
This also allows us to change the colours for warnings or errors
easily by updating just a single line of code, if ever desired.
It also permits us to call `pywikibot.warning()` and
`pywikibot.error()` resp. instead of `pywikibot.stdout()` --
that would make much sense, but before we do so we must investigate
whether the modification requires any changes to the setup for
the logs on the server, so this is postponed to a separate commit.

At least for now, we do not use a separate function for subheadings,
but just replace the `color` parameter of the `out()` function
by a simple optional boolean `heading` parameter.  The emphasis
for subheadings is just an emphasis and semantically not as special
as errors and warnings; just like normal info output, it will always
use the default `pywikibot` output function; therefore it would
probably be just too much to define a separate function for it.

As a side effect, we rename the variable `error` which is used by
some `except` exception handlers, because we now need it for the new
`error()` function.  Let’s utilize the very common name `exc` instead.
Actually the object referenced by the variable is an exception object,
not an error ;–), so the new name is also even more appropriate.