Make Candidate.cleanTitle() consider the alternative
We use `Candidate.cleanTitle()` to get a neat and tidy title
of the nominated image, for use in gallery page captions etc.
It derives that title from the name of the nomination subpage.
But for FP nominations which contain one or more ‘alternative’
images, the result can be wrong: the name of the nomination subpage
usually refers to the originally nominated image, however
if an alternative image is the chosen variant, the clean title
must be based on the name of the selected alternative file.

Therefore let’s make `cleanTitle()` consider the alternative
filename if there is one.  For now the method still defaults
to the old behaviour and the use of the alternative name must
be enabled via the ‘alternative’ parameter; but this can be changed
easily in the future after a careful inspection of all places
in the code which call this method.