Fix caption for featured alts in addToCurrentMonth()
For FP nominations which contain one or more ‘alternative’ images,
the caption of the entry created by `Candidate.addToCurrentMonth()`
was often wrong.  It was always based on the name of the
nomination subpage and therefore on the name of the originally
nominated image.  However if an alternative image is the chosen
variant, the caption must be based on the name of the selected
alternative file.  This commit fixes the caption in such cases.