Improve Candidate.moveToLog()
1. The regex which is used to remove the nomination subpage
inclusion from the list of current candidates ends with
`}}.*?\n?`.  This means that it removes also any stuff which
occurs between the inclusion `{{...}}` and the end of the line.
E.g., if a user puts a nomination inclusion on the same line
right after another one, like in `{{...}} {{...}}`,
and the first nomination is closed, our bot program removes
not only the first inclusion, but also the 2nd one with it.
That’s fatal, and such things can actually happen (sometimes
users make really strange entries on the candidate list).
Therefore let’s make that regex stricter.  Of course we can
tolerate some harmless spaces before or after the `{{...}}`,
but there should be nothing else on the same line, else we
must not remove the entry in order to avoid any damage.
And, well, we better escape the opening and closing braces
in the regex because braces have a special regex meaning, too,
so future versions of Python’s regex engine could fail
because the regex contains unescaped braces.

2. In the commit message used when changing the log page
or the list of candidates, we link to the nominated image.
But this is a bit misleading because this method does not
add the image to the log nor does it remove the image
from the list of candidates, but it adds the nomination
subpage to the log and removes it from the candidate list.
Especially for nominations containing an alternative and
for set nominations this is a big difference.
Therefore let’s link to the nomination subpage itself
so that if an user inspects the history of the log page
or of the candidate list and clicks on the link the user
is immediately directed to the nomination subpage.

3. On the same lines, let’s make the messages printed to the
console or log refer to the name of the nomination subpage
instead of to the name of the nominated image file;
after all these messages refer to issues with the nomination
subpage, not to issues with the image file.

4. Most methods which are part of the parking procedure
and change the contents of a page on Wikimedia Commons
use commit messages in the simple past, like ‘Added (filename)’.
But the method `Candidate.moveToLog()` uses the present
progressive and says ‘Adding (name)’ and ‘Removing (name)’.
Now while the present progressive is very common in status
messages printed by this bot program and other software,
it is rather unusual in commit messages for Wiki projects;
most people use the simple past.  Therefore let’s bring also
the commit messages produced by `moveToLog()` into line
with messages from other parts of the parking procedure and
with messages by most Wiki users and use ‘Added (name)’ and
‘Removed (name)’ resp.

5. Let’s use local variables for stuff we need more than once,
also in order to make the method even easier to understand.

6. Let’s remove a misleading FIXME which is completely obsolete
and add a helpful explanation instead.

7. Let’s improve the structure of the method and especially
the structural comments.