Improve output formatting of Candidate.printAllInfo()
Let’s make the formatting of the output printed by the method
`Candidate.printAllInfo()` more consistent:

1. Print the count of sections with at least two figures and
a leading zero (if necessary), just like all similar entries.

2. Always use 5 letters for the output of `isWithdrawn()`.
Until now `True` used 4, `False` used 5 letters, leading to
an irritating misalignment of the last column of the whole table
if there is at least one withdrawn candidate in the list.
To achieve a consistent alignment of the last, very important
column, we must always print the penultimate column with
the same count of characters.  (NB: We could also use
`str(self.isWithdrawn()).ljust(5)`, but that depends on the
length of the longer value, so IMHO it is even clearer to use
the simple `"True " if self.isWithdrawn() else "False"`.)

At this opportunity, let’s also improve the docstring.