Fix and improve Candidate.notifyUploader()
1. In the method `Candidate.notifyUploader()` we test whether
the promotion template for a new FP is already present
on the uploader’s talk page or not, and add the template
only if the template is not present yet.  But this test fails
because we search the contents of the talk page for the wrong
template: we search for `{{FPpromotion}}`, but the method
adds `{{FPpromotedUploader}}`.  Therefore if we ever repeat
the parking procedure for the same nomination (e.g. because
the previous run has been interrupted), we will add another
promotion template for the same FP.  Let’s fix this and also
avoid any other possible differences between the test and
the wikitext actually added by the method: first assemble
the complete text of the template with its parameters,
then use it both for the search and for adding it (if needed)
to the wikitext of the page.

2. The same method sometimes exits too early, without handling
all image files from a set nomination.  If the talk page of an
uploader is undefined or if it already contains the promotion
template for one of the images from the set, we call `return`.
But this means that any other images from a set nomination
are not processed anymore; and that’s not good because
if the parking procedure has previously been interrupted
it is possible that some of the images have been handled by
the method, but some others have not been handled.
Therefore let’s fix both instances by calling `continue`
instead of `return`, so we skip just the current image and
continue with the next image from the set nomination.

3. Print a hint to the CLI resp. to the log that we skip a file
if the user who has uploaded it is identical with the nominator.
In all similar situations we are adding a message, so it was
rather surprising that we did not do it in this case.

4. Avoid to report more than once that the talk page of a certain
uploader is undefined or locked.

5. Get rid of repeated calls of `Candidate.nominator()` for one
and the same nomination as well as of double calls of `uploader()`
for one and the same file.

6. There is no need for an underscore in the ‘User talk:’ prefix,
we can just use a simple space (looks much better in messages).

7. At this opportunity, let’s improve the messages which are
printed if something is wrong to make them clearer and more
helpful, and also the comments on these lines of code.

8. In the method docstring, emphasize that and why this method
must handle the files one by one.  This seems trivial, of course,
but it is the key for understanding the method and how it differs
from `Candidate.notifyNominator()`.