Improve skipping of image and 'Skipping...' messages
Sometimes the method `Candidate.addAssessments()` exits too early,
without handling all image files from a set nomination.
If the description page of an image already contains an identical
`{{Assessments}}` template, we call `return`.  But this means
that any other images from a set nomination are not processed
anymore; and that’s not good because if the parking procedure
has previously been interrupted it is possible that some of
the images have been handled by the method, but some others not.
Let’s fix this by calling `continue` instead of `return`,
so we skip just the current image and continue with the next one.

Improve the messages which are printed to the CLI and to the log
if during the parking procedure a method, a certain image,
a talk page, etc. is skipped:

1. by referring to the concrete filename if the issue is related
to an individual image file;

2. to the name of the nomination subpage if the issue is related
to the complete nomination;

3. to the name of the user talk page if the issue is related
to the talk page of the nominator; etc.

Improve the wording of all these messages to make them more helpful
and improve also some related comments.