Improve Candidate.addToCurrentMonth()
1. The initial search for the voting results can reuse the constant
`FinalVotesR` instead of defining yet another regex; and we don’t
need to iterate over its matches because in `Candidate.park()`
we have already checked that there must be exactly a single result.

2. If that search does not find a verified result, we must not
continue and add an invalid entry (as the old code used to do);
better print a clear error message and skip the rest of the method.

3. The flow of control in the middle part of the method was
a bit unclear and quite prone to misunderstandings.
Therefore reworked order and structure of that part, improved
the comments, error messages, etc.

4. The name of a set nomination was emphasized as bold text,
but that did not work well because the name is inserted into
a span of bold text later.  As a result set entries showed
an odd combination of bold plus italic fonts plus typewriter-like
quotation marks, one of them in italics, the other one not.
Let’s use a simple and clear notation instead, starting with
‘Set: ...’, which allows to find set nominations more easily
(suggested by Wikimedia Commons user W.carter, thanks!).

5. Every entry contains a link to the new FP (just click on the
thumbnail image), but it was missing a link to the nomination.
Let’s add it, at least for now in a rather discrete way,
by linking the number of the entry to the nomination subpage.

6. Assembling the final entry with the good old `%` operator shows
its limits: it has no less than 8 arguments here and it is really
hard to see which `%s` will be replaced by what value.  Python’s
f-strings make the code much clearer and easier to read.

7. Let’s improve the commit message for set nominations:
instead of mentioning just the name of the first file from the set
(which makes the comment look like a normal single-file nomination),
mention and link to the set nomination.

8. Let’s improve the comments.  Especially explain why even a new,
completely empty page yields an entry count of 1.