Post stack trace of parking procedure error to FPC talk
Right now we have the unpleasant situation that FPCBot stops
right after step 3 of the parking procedure when processing
any successful FPC nomination.  Probably this is related to
my latest changes to step 4, namely to the changes in the code
of `Candidate.addToCurrentMonth()` in the preceding commit.
I am very sorry for this and any inconveniences it may cause.
But the real problem is that I cannot reproduce the issue
at all when running the program locally -- everything seems
to work perfectly fine then.  At the same time I have no access
to the log files of FPCBot on the server.  This makes debugging
like looking for a needle in the haystack.

I could just revert my latest changes to the code, but that
would be a lost chance.  We need to understand what exactly
is not working, only then we can be sure to avoid the same
error in the future.  So before going back let’s try to go
one more step forward.

As kind of a poor man’s log file, I deploy some code I wanted
to commit later, namely a simple function to post a message
to the FPC talk page.  Let’s use it to post a stack trace
to that talk page if there is an uncaught exception during
the parking procedure.  It this works as intended (and it does
work very well during local tests), it allows us to catch
and hence to fix exceptions in the parking procedure without
having to fight for freeing the log files ;–).  And we can
re-use this code later for more regular issues, e.g. to post
a message to the talk page if FPCBot cannot sort a new FP
into the appropriate gallery section, so that human editors
should step in and fix this; because this was what I had
in mind when developing the `ask_for_help()` function.