Don't use 'count' in str.replace() as keyword argument
Since Python 3.13, we can use `str.replace()` with `count=...`
as a keyword argument; and this makes sense when we consider
that similar methods like `re.sub()` even require us to pass
the max. count of replacements as a keyword argument.
But older Python versions don’t support this in `str.replace()`
and require us to pass the count as an plain integer.

This also explains why FPCBot did stop to work after my last
commit in spite of the fact that I had tested that commit really
thoroughly -- my machine runs Python 3.13, but the Wikimedia labs
server seems to run FPCBot with some older version of Python.

Let’s fix this issue simply by passing the count as a simple
integer argument.  At this opportunity let’s also add the count
to yet another `str.replace()` call which needs to replace
max. one instance of the substring, just like in the other cases.