Format stack trace with <pre>
In the recently added code to print a stack trace for uncaught
exceptions during the parking procedure, format the stack trace
with `<pre>...</pre>` instead of simply adding a space at the
beginning of every line.  Using `<pre>` fixes a problem in the
display of the stack trace (when using simple spaces to mark
preformatted text, as we did before, the Mediawiki parser still
interprets `~~~~` as a signature with timestamp and resolves it).
The new code fixes also yet another nasty problem with older
Python versions on the server: they do not allow backslashes
in the expression part of f-strings (SyntaxError).