Move uncaught exception reporting to checkCandidates()
If we retain the idea to report uncaught exceptions (which stop
the bot program, of course) on the FPC talk page, at least for now,
it makes more sense to move the code two levels up to the function
`checkCandidates()`.  In this function the code integrates well
into the existing `try ... except ...` clauses, and here it will
report uncaught exceptions not only from the parking procedure,
but from all major tasks of the bot program.

At this opportunity, improve the wording of the error message
by including a hint to the offending nomination -- this way
it is much more informative also for ‘normal’ FPC regulars --,
and polish the whole function a little bit.