Improve error handling and error message formatting
This commit improves the error handling and related points in
various methods and functions.  These changes are also required
as preliminary work for the next commit which introduces
requests for help on the FPC talk page.

1. If `findCandidates()` can’t read the list of candidates,
we must print a useful error message and return.

2. If `findCandidates()` can’t find a single nomination subpage
in the list, we must print a useful error message and return.
There is no point in analysing an empty list.

3. In `addToGalleryPage()` we should check whether the gallery
page specified by the gallery link exists (it’s not uncommon
that people make errors when adding the gallery link)
and generate a useful error message if this is not the case.
Likewise, because this is a very important task, let’s also
generate a specific error message when reading the page fails.

4. If `handlePassedCandidate()` does not find the file which was
specified as selected alternative version, we printed a message,
but forgot to cancel the parking procedure.  We clearly must add
a `return` here, just like in all similar tests.  Furthermore
the message which is printed in this case can be improved:
it did start with the name of the alternative, but that’s a bit
confusing; better start the message (like all other ones)
with the cut title of the nomination and mention the faulty
alternative value only later.

5. Generalized the exception handler in `findCandidates()`
to catch all page-related Pywikibot exceptions.

6. The messages printed (i) if a nomination seems to contain
several verified results, (ii) if the value of the `featured`
parameter is unknown, (iii) if the gallery link is undefined,
(iv) if the alternative was not found, or (v) if no promoted
image files are found in the nomination ... are all clear
error messages, so we should print them using our usual
`error()` function, not with the normal `out()` function.

7. We need to improve the error message which is printed
if `park()` notices that the success value from the results
is neither `yes` nor `no`: the status is not ‘unknown’,
but ‘invalid’, and because `park()` is also called for
delisting candidates, it is not a ‘feature status’, but
rather a success value.

8. In `findCandidates()` it makes much more sense to re-use
the local variable `subpage_name` than to call the method
`subpage.title()`.  This also makes the code clearer.

9. In `addToGalleryPage()` we should issue a warning if the
gallery section is empty or invalid and FPs are sorted
into the ‘Unsorted’ section.  That’s quite common, but still
worth a warning because it requires extra work.

10. Improved comments and code formatting by the way,
to make the code even clearer and easier to understand.