Improve parameter name in addToGalleryPage()
Renamed the parameter `gallery` of the method
`Candidate.addToGalleryPage()` to `gallery_link`.
This makes clear that the parameter does not represent
the gallery page etc., but just the original gallery link
from the nomination subpage.

BTW: Dropped the line `assert gallery_link and files` from the
same method.  The code which calls this method already checks
both values, so the assertion seems a bit too much.
If you remove these checks or call the method in another way,
well, then it’s up to you to make sure that the values
of both parameters are valid.