Report common problems on the FPC talk page
When the bot cannot accomplish one of its tasks, the reason
is often just a little typo in a gallery link, a formatting
error in a nomination, a renamed file or nomination subpage,
etc. -- i.e., little issues which a human editor can fix easily,
but which stop the bot doing its duties or halt it completely.
Of course the log files contain all information which is needed
to fix these problems, but even if we would make the log files
available for everybody people would have a hard time to find
the crucial information for solving their problem.

But since a few commits we have the function `ask_for_help()`,
so we can make good use of it.  Let’s just post a short notice
on the FPC talk page whenever the bot needs the help of human
editors.  I have discussed this with the experienced Wikimedia
Commons user W.carter.  In the last year W.carter and me
have been the main maintainers of the FPC gallery pages,
categories and related stuff.  There are more people who want
to help with the FP stuff, but often they miss need clear
information about what must be done, why and were.  Therefore
we hope that letting the bot post concise requests for help
will enable more people to take part in the maintenance
for the FP gallery pages etc.

This commit implements that idea by applying `ask_for_help()`
in many places where the bot needs the aid of human editors.
Of course some messages may be missing, some messages may be
considered unnecessary (because it is unlikely that they will
be needed often), and certainly the wording of some messages
could be improved.  But most of the work is done.  Let’s see
how it works and improve and polish minor points later.