Improve message about invalid gallery section anchor
The message that the section anchor of a gallery link is missing
or invalid will probably be the most frequently used message
of the bot.  The reason is simply that nominators often do not
fully understand the purpose and structure of a gallery link.
Therefore we better provide a longer message which also explains
the conditions a section anchor must meet in order to be valid.