Abbreviate file paths to filenames in stack trace
The stack trace which we print if the bot program fails
with an uncaught exception should be a simple as possible.
Therefore let’s abbreviate the file paths which are included
before every single entry of the stack trace to filenames.
Most often they will just point to our own program,
and even in the rare other cases they will be about a module
from the Python Standard Library or from Pywikibot;
in all cases we can easily recognize the file with the help
of the mere filename. This also avoids to expose the directory
structure of the server; people still believing in security
by obscurity could consider that as a security risk.