Use 'FP promotion' in edit summaries
In all the subheadings, templates, etc. which are used and
generated during the promotion of a new featured picture,
we use the phrase ‘FP promotion’.  The only exception are
the edit summaries which are used by the bot when saving
its changes: three of them say ‘FPC promotion of ...’.
It seems more appropriate to use ‘FP promotion of ...’ here,
too.  Each file which is handled in this process is already
a FP candidate (FPC), hence it is not promoted to a FPC,
but to full FP status now.