Fix, improve, rename Candidate.fixHeader()
The previous implementation of `Candidate.fixHeader()` added
the ‘(not) featured’, ‘(not) delisted’ keyword a second time
to the nomination heading if there is ‘something’ before
the heading, e.g. whitespace (this is not even that rare,
quite some nomination subpages start with an empty line etc.).

This commit fixes the test, removes the irritating leading
whitespace and improves the docstring, the comments and the
code structure.  It also renames the method to `fixHeading()`
because we are not dealing with a header here (‘header’ makes
one think of the HTML `<head>` section and similar complex
document structures), but just with the simple heading of
a nomination subpage.

Some nomination subpages even start with complete cruft like
`{{DEFAULTSORT|}}`, inserted by accident during voting [1].
To catch such really irritating aberrations, and also to find
nominations in which the heading is misformatted or damaged,
we add a more rigorous test.  If it fails, we print a warning
to the CLI and ask for help with our usual function.

[1] For example, see the history of this nomination:
https://commons.wikimedia.org/wiki/Commons:Featured_picture_candidates/
File:Pez_an%C3%A9mona_ensillado_(Amphiprion_polymnus)_en_una_
an%C3%A9mona_alfombra_de_Mertens_(Stichodactyla_mertensii),_
Anilao,_Filipinas,_2023-08-23,_DD_108.jpg