Improve getResultString(), add gallery in alternatives
When the bot program closes a FP nomination with alternative
images, it does not count the votes and puts just `X` into
the vote count parameters of the `{{FPC-results-unreviewed}}`
template.  But the bot should copy the gallery link into the
`gallery` parameter: this value does not depend at all
on the selected alternative and copying it saves the reviewer
from copying it (and adding typos to the value BTW ;–).
Therefore always fill the `gallery` parameter of the template.

In nominations with several alternatives, we use `X` as the
placeholder value for the `support`, `oppose`, and `neutral`
parameters, but `no` as placeholder value for the `featured`
parameter.  This is a bit misleading because it seems to
suggest that the bot has counted the votes and concluded
that no alternative is featured.  Therefore better use `X`
as placeholder value for `featured`, too.

At this opportunity, let’s tidy up both `getResultString()`
implementations.  The wording of the hint “the alternative
needs to be specified” was improved with the help of
Wikimedia Commons user Cmao20 (thanks!).  Added docstrings
to the methods and moved the insertion of the leading `\n\n`
to `closePage()` because it is the same for all three variants
of the template.  IMHO f-strings make the code much clearer
than the old `%` operator: with f-strings we can see at once
which value is assigned to which one of the six or seven
template parameters.

I have tested the new code by comparing the results with the
results from the old code for all nominations from January
to July 2025 and the results are identical -- of course
apart from the `gallery` parameter and from the wording of
the hint which is added to nominations with alternatives.