Improve wording of message in getResultString()
Wikimedia Commons user Cmao20 has pointed out that if we include
a hint in the `{{FPC-results-unreviewed}}` template, it must be
clear and comprehensible especially for less experienced users
(the nerds do not need the hint at all, of course).  Therefore
we better avoid technical terms like ‘parameter’.  Let’s use
the wording proposed by Cmao20 (thanks!), it is a good compromise
between the need to be brief and the need to be as comprehensible
as possible.