Identify, notify and mention original creator
Sometimes the original creator of a Featured picture is different
from the uploader, e.g. when we promote a variant of an image
which has been retouched by another user.  In this case we should
notify also the original creator: credit where credit is due.

Of course there is no generally applicable and reliable way
to determine the original creator of an image; therefore the bot
has never notified creators.  (The source code contained several
hints about a planned creator notification, though it has never
been implemented.)

But we can offer a simple workaround which should work well in
most cases.  If the original creator is different from uploader
and nominator, nominators should use the phrase

    {{Info}} ... created by [[User:...|...]]

on the nomination subpage in identify the original creator.
This phrase is already used in the current default text for new
FP nominations.  We also support the common variant ‘created and
(adverb?) uploaded by’.  If this phrase is present, we can identify
the original creator and notify the user about the FP promotion.

Likewise we also mention the original creator in the entry for
the monthly overview of new FPs if (and only if) we can identify
the creator and if the creator is different from the uploader.
(This means we mention the creator if the creator is identical
to the nominator, but different from the uploader.  In most cases
the uploader is also the creator, so we should acknowledge the
creator separately if the uploader is not the creator.)