Cache creator, uploader, nominator names
Retrieving the user name of the creator, uploader or nominator
can take some time because it requires to read a Commons page
or to request revision data via the Mediawiki API.  We are not
in a hurry, of course, but doing this twice (because we are
referring to each of these user names in two different methods)
and potentially even more often in the future (if we want to use
the usernames for additional functionality) is irksome.
Therefore cache these names in three new properties of the
`Candidate` class and just return these values whenever we call
again `Candidate.nominator()` etc.

This also gives opportunity to tidy up these methods and to split
out shared functionality as new functions.  For example, it doesn’t
make much sense that `Candidate.nominator()` and `uploader()` test
the value of `history` only *after* using that value; and it is
a bit inelegant that both of them iterate over an iterator with
a single element instead of getting that element directly.
We can make both subroutines clearer and safer with the help of
functions which bundle the shared code, by using the Pywikibot
`Page.oldest_revision` property and by catching related exceptions
properly.