Pass candidate list page name to class initializers
The name of the Commons page which includes all candidates and
which therefore acts as the source for all operations is stored
in an instance variable of the `Candidate` class and its subclasses.
But it was not passed as a parameter to the initializers
of these classes (as it is the case with all similar values),
but hard-coded at the end of the subclass initializers.

Remove this idiosyncrasy, pass the list page name as parameter
to the initializers of the classes and assign it there to the
instance variable.  This way the value which is defined as a
local constant in the `main()` function is passed through all
functions -- `findCandidates()`, `checkCandidates()` to the
class instances.  So we make sure that one and the same value
is used in all these places and get rid of the repeated
definition of one and the same constant value in three places.

This also fixes a little bug.  If there was an error during
the `-test` task (recounting old nominations from a log);
or if one changed the value of `candidates_page` in `main()`
in order to test the bot program with archived nominations
instead of the current ones, and an error occurred;
the bot would still have reported that there were problems
processing nominations from the default candidate list,
because for error reporting it used the hard-coded list page
name from the `__init__()` methods of the two subclasses.
This was irritating and misleading, and is fixed now.

Remove some related variable/parameter names and remove the
unnecessary and slightly irritating underscores from the initial
list page name definitions in `main()`.