Tidy up imageCount()
In the method `Candidate.imageCount()`, do not re-count the images
if the result of the previous counting was 0 (we can achieve this
simply by testing `if self._imgCount is not None`).

Get rid of the pointless iterator `re.finditer(...)`.  We should
either iterate over the matches one by one or get them all at once
with `re.Pattern.findall()`; but it makes no sense to use an iterator
to append each single match to a list.  Using `ImagesR.findall()`
means that instead of `re.Match` objects we get get a list of strings
or of tuples of strings, corresponding to the capturing groups
in the regex.  To access the complete image link we have to add
another group to `ImagesR` and therefore must update the code in
`Candidate.fileName()` accordingly.

Improve the docstring, the names of variables, etc., and remove
the blank lines (not needed in such a short method).

In the related method `Candidate.isIgnored()`, we should also
consider the possibility that `imageCount()` returns 0;
this is possible, of course, if something is wrong with formatting
etc. of the nomination subpage, and we should flag any nomination
with an image count of 0 for manual counting, too.